pyproject.toml
# Moon Dev AI Trading Agents - UV Project Configuration
# Built with ❤️ by Moon Dev

[project]
name = "moon-dev-ai-agents"
version = "2.0.0"
description = "Experimental AI trading system with 48+ specialized agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Moon Dev", email = "moondev@example.com"}
]

keywords = ["trading", "ai", "agents", "crypto", "solana", "hyperliquid"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core AI/LLM Dependencies
    "anthropic>=0.40.0",
    "openai>=1.59.5",
    "groq>=0.16.0",
    "google-generativeai>=0.8.3",

    # Crypto/Trading
    "solana>=0.35.1",
    "solders>=0.21.0",
    "hyperliquid-python-sdk>=0.20.0",
    "eth-account>=0.11.0",
    "web3>=6.20.1",

    # Data & Analysis
    "pandas>=2.1.4",
    "numpy>=1.26.2",
    "pandas-ta>=0.3.14b0",
    "TA-Lib>=0.4.32",
    "Backtesting>=0.3.3",
    "yfinance>=0.2.43",

    # Utilities
    "python-dotenv>=1.0.0",
    "termcolor>=2.3.0",
    "colorama>=0.4.6",
    "requests>=2.31.0",
    "tqdm>=4.67.1",
    "schedule>=1.2.0",
    "psutil>=5.9.8",
    "httpx>=0.28.1",

    # Media Processing
    "opencv-python>=4.10.0.84",
    "pillow>=10.3.0",
    "ffmpeg-python>=0.2.0",
    "whisper>=1.1.10",
    "openai-whisper>=20231117",
    "elevenlabs>=0.2.24",

    # Document Processing
    "PyPDF2>=3.0.0",
    "youtube-transcript-api>=0.6.2",

    # Social Media
    "twikit>=2.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://moondev.com"
Documentation = "https://github.com/moon-dev/ai-agents"
Repository = "https://github.com/moon-dev/ai-agents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
